Fix feature style and improve readme
* Make the debugger attribute of the interpreter compatible with the
  format proposed by the profiler that avoids using a mock
  implementation.

* Modularize the profiler and split the feature requirements defined by
  `profiler-any` and `profiler-gas` with their respective modules.

* Avoid importing internal preludes related to the library to improve
  source readability.

* Include a changelog.
