+++ CHANGELOG.md
@@ -1,0 +1,24 @@
+# CHANGELOG
+
+## Pending
+
+### Breaking changes
+
+### Features
+
+### Improvements
+
+- [#65] Fix feature style and improve readme.
+
+### Bugfixes
+
+## v0.1.1
+
+### Improvements
+
+- [#64] Fix license SPDX identifier.
+
+## v0.1.0 (Initial release of FuelLabs/fuel-vm)
+
+[#65]: https://github.com/FuelLabs/fuel-vm/pull/65
+[#64]: https://github.com/FuelLabs/fuel-vm/pull/64
